fix: separação dos controllers em arquivos
