+++ backend/src/controllers/User/updateUser.js
@@ -1,0 +1,24 @@
+import User from '../../models/User.js'
+import formatUser from '../../utils/formatUser.js'
+
+export async function updateUser(req, res) {
+    const { id } = req.params
+    const { name, bio } = req.body
+
+    try {
+        const updatedUser = await User.findByIdAndUpdate(
+            id,
+            { name, bio },
+            { new: true }
+        )
+
+        if (!updatedUser) {
+            return res.status(404).json({ res: 'Usuário não encontrado' })
+        }
+
+        return res.status(200).json(formatUser(updatedUser))
+
+    } catch (error) {
+        return res.status(500).json({ res: 'Erro ao atualizar usuário', error: error.message })
+    }
+}
